# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CBT API Documentation",
        "description": "API Documentation for Computer Based Test (CBT) System",
        "contact": {
            "email": "admin@cbt.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/admin/nilai": {
            "post": {
                "tags": [
                    "Admin - Nilai"
                ],
                "summary": "Simpan nilai siswa",
                "operationId": "f517eb92331da2c72e0d1b88c30dce5c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "mapel_id",
                                    "skor"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "mapel_id": {
                                        "type": "integer"
                                    },
                                    "skor": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/nilai/{user_id}": {
            "get": {
                "tags": [
                    "Admin - Nilai"
                ],
                "summary": "Lihat nilai per siswa",
                "operationId": "7694e3348c6186e872cd9b5265e16a5a",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/soal/{mapel_id}": {
            "get": {
                "tags": [
                    "Admin - Soal"
                ],
                "summary": "Daftar semua soal per mata pelajaran",
                "operationId": "dc40ebcaa0926aee8c155ba6888912b0",
                "parameters": [
                    {
                        "name": "mapel_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/soal/import": {
            "post": {
                "tags": [
                    "Admin - Soal"
                ],
                "summary": "Import soal dari Excel (Soon)",
                "operationId": "ee869dc718aeee5ed3d36bcb71677902",
                "responses": {
                    "501": {
                        "description": "Not Implemented"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/soal/{id}": {
            "put": {
                "tags": [
                    "Admin - Soal"
                ],
                "summary": "Update soal",
                "operationId": "85a12420a7106200d2eb8aac6e36ce85",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "pertanyaan": {
                                        "type": "string"
                                    },
                                    "opsi_a": {
                                        "type": "string"
                                    },
                                    "opsi_b": {
                                        "type": "string"
                                    },
                                    "opsi_c": {
                                        "type": "string"
                                    },
                                    "opsi_d": {
                                        "type": "string"
                                    },
                                    "jawaban_benar": {
                                        "type": "string",
                                        "enum": [
                                            "A",
                                            "B",
                                            "C",
                                            "D"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Soal"
                ],
                "summary": "Hapus soal",
                "operationId": "527e183435b0cb5781e3dd6ad315e946",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}